# faster simple path filling