before:
```
[pixie] rect opaque ................ 0.060 ms      0.062 ms    ±0.002  x1000
[pixie] rect not opaque ............ 0.145 ms      0.147 ms    ±0.002  x1000
[pixie] roundedRect opaque ......... 0.102 ms      0.104 ms    ±0.002  x1000
[pixie] roundedRect not opaque ..... 0.255 ms      0.258 ms    ±0.005  x1000
[pixie] tiger ..................... 10.053 ms     10.265 ms    ±0.211   x486
```

after:
```
[pixie] rect opaque ................ 0.046 ms      0.055 ms    ±0.010  x1000
[pixie] rect not opaque ............ 0.127 ms      0.129 ms    ±0.005  x1000
[pixie] roundedRect opaque ......... 0.091 ms      0.093 ms    ±0.002  x1000
[pixie] roundedRect not opaque ..... 0.239 ms      0.242 ms    ±0.005  x1000
[pixie] tiger ..................... 10.086 ms     10.141 ms    ±0.061   x491
```

----

cairo:
```
[cairo] rect opaque ................ 0.035 ms      0.038 ms    ±0.005  x1000
[cairo] rect not opaque ............ 0.142 ms      0.145 ms    ±0.006  x1000 < pixie is now winning here
[cairo] roundedRect opaque ......... 0.053 ms      0.055 ms    ±0.003  x1000
[cairo] roundedRect not opaque ..... 0.192 ms      0.194 ms    ±0.003  x1000
[cairo] tiger ..................... 12.885 ms     13.009 ms    ±0.137   x383 < pixie has won here for a while now
```